mensajes de validación de información faltante

+++ src/Components/ClientForms/DeleteClient.js
@@ -2,6 +2,7 @@ import React, { useState } from "react";
 import './FormsClient.css'
 import { getClientByPhone } from "../../services/clientsService";
 import $ from "jquery";
+import Swal from "sweetalert2";
 
 export default function DeleteClient(props){
 
@@ -31,11 +32,10 @@ export default function DeleteClient(props){
             
             if (response.status === 200)
             {
-                console.log(response.data);
                 let name = response.data.name.split(" ");
-                document.getElementById("name").value = name[0];
-                document.getElementById("lastname").value = name[1];
-                document.getElementById("phone").value = response.data.phone_number;
+                setName(name[0]);
+                setLastname(name[1]);
+                setPhone(response.data.phone_number);
                 $( ".enabling" ).prop( "disabled", false );
                 $( ".hidden" ).prop( "hidden", false );
                 setId(response.data.id);
@@ -53,8 +53,25 @@ export default function DeleteClient(props){
 
         const handleFormSubmit = (e) => {
             e.preventDefault();
-    
-            props.onDeleteClient(Id, Name, Lastname, Phone);
+
+            Swal.fire({
+                title: '¿Está segura de modificar clienta?',
+                icon: 'question',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Sí, continuar',
+                cancelButtonText: 'Regresar'
+              }).then((result) => {
+                if(result.isConfirmed){
+                    props.onDeleteClient(Id, Name, Lastname, Phone);
+                    Swal.fire({
+                        title: 'Edición exitosa',
+                        icon: 'success',
+                      })
+                }
+              })
+            
         }
 
     return(
